server/requirements.txt
MySQL-python==1.2.5
Django==1.10.5
djangorestframework=3.5.4
ipython==5.2.2
django-cors-headers==2.0.2
